Migrate code from road-core/rag-content
This commit migrates code from road-core/rag-content into this
repository. Only code which is related to the generation of the vector
database is being migrated.

Code related to any specific product consuming rag-content is
intentionaly left out of this migration as each consumer of the
rag-content should create their own repositories [1].

[1] https://github.com/road-core/rag-content