Remove references to OpenShift
Remove references to OpenShift both from README.md and Makefile. This
commit updates the README.md so that it does not contain any reference
to OpenShift and so that it contains an example of building a dummy
vector database.

This way, the README.md still provides information about how to build
the
vector database, but it is not related to any specific service.