Llama-stack register DBs script
When bringing the vector-io file with our knowledge to a new
host/container, configuring the vector-io provider, and running
llama-stack, the service doesn't automatically register the vector-dbs
present in the file; we'll need to manually register them.

This patch adds a sample script called `register-llama-dbs.py` that can
do that for us automatically.

The idea is to either run that script when building the container for
llama-stack (if we include the knowledge in it) or have an init
container in our pod that runs it.