Make document processor compatible with latest llama-stack
In llama-stack version 0.2.13 onwards the default mechanism of adding
data (manual chunking) will result in a database that cannot be consumed
without an exception:

```
  File "/home/ec2-user/rag-content/venv/lib64/python3.12/site-packages/llama_stack/providers/inline/tool_runtime/rag/memory.py", line 173, in query
    metadata_for_context[k] = getattr(chunk.chunk_metadata, k)
                              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
AttributeError: 'NoneType' object has no attribute 'chunk_id'
```

This error is caused by a new patch [1] that separates between the
metadata used for training (`metadata`) and the metadata needed for
llama-stack's operation (`chunk_metadata`).

This repository currently only supports python 3.11 which is not
supported by any llama-stack version that breaks our current code, but
in this patch we are adding compatibility code to support both the
currently used and newer versions in order to facilitate moving this
project to newer python releases without complications.

[1]: https://github.com/meta-llama/llama-stack/commit/82f13fe83e5c9c802595b05fb2bc2ef4ddedbe3c