Ignore docstrings missing in test files
Currently ruff fails if a __init__.py or a .py file in our tests is
missing a docstring for the whole file.  This usually makes no sense,
because the purpose of a unit test file is usually clear, and the test
class still has a docstring explaining things.

This patch removes the need for those docstrings. To do so it also moves
the command line `--per-file-ignores` argument in the `Makefile`'s calls
ro ruff to the `pyproject.toml` so it doesn't override the new additions
replacing them.

It also prevents some pydocstyle conflicting errors. This means that are
checks that cannot be enabled at the same time because they are
contradictory and therefore cannot be satisfied simultaneously, such as:

- D203 wants 1 blank line before class docstring, but D211 want's not
  blank lines.

- D212 wants multi-line docstring summary to start at the first time,
  but D213 wants it to start on the second line.