doc processor: Markdown skip header-only chunks
When processing documents using the markdown chunker we may end up with
chunks that don't really have any useful information, as in having only
headers.

For example, the following text:

    # Configuration

    ## cinder.conf

    Blah, blah

Will result in 2 chunks:

    # Configuration

and

    ## cinder.conf

    Blah, blah

With the first chunk being useless, moreover considering the second
chunk has this in its metadata (that is used for the embedding):

    header_path: /Configuration

In this patch we change how we process markdown and html doc types and
remove chunks that have no real content. We consider chunks as empty
when they only contain:

- Empty lines
- Empty code blocks
- Horizontal rules
- ATX header lines
- Setext header lines

Closes: LCORE-1491