doc processor: Markdown skip only header chunks
When processing documents using the markdown chunker we may end up with
chunks that don't really have any useful information, as in having only
headers.

For example, the following text:

    # Configuration

    ## cinder.conf

    Blah, blah

Will result in 2 chunks:

    # Configuration

and

    ## cinder.conf

    Blah, blah

With the first chunk being useless, moreover considering the second
chunk has this in its metadata (that is used for the embedding):

    header_path: /Configuration

In this patch we change how we process markdown doc types and remove
chunks that have no real content because they only have headers or only
have space related characters.